Barang keluar, login, data admin

+++ app/Config/Routes.php
@@ -35,48 +35,55 @@ $routes->set404Override();
 
 // We get a performance increase by specifying the default
 // route since we don't have to scan directories.
+$routes->get('/', 'home::index', ['filter' => 'auth']);
+
+// Login
+$routes->get('/login', 'Login::index');
+$routes->post('/login/auth', 'Login::auth');
+$routes->get('/logout', 'Login::logout');
 
 // Data Admin
-$routes->get('/data_admin', 'Admin::index');
+$routes->get('/data_admin', 'Admin::index', ['filter' => 'auth']);
+$routes->get('/data_admin/create', 'Admin::create', ['filter' => 'auth']);
+$routes->post('/data_admin/save', 'Admin::save', ['filter' => 'auth']);
 
 
 // Jenis Barang
-$routes->get('/', 'home::index');
-$routes->get('/jenisbarang', 'Kategori::index');
-$routes->post('/jenisbarang', 'Kategori::store');
-$routes->get('/jenisbarang/edit/(:any)', 'Kategori::edit/$1');
-$routes->put('/jenisbarang/(:any)', 'Kategori::update/$1');
-$routes->get('/jenisbarang/create', 'Kategori::create');
-$routes->delete('/jenisbarang/(:any)', 'Kategori::destroy/$1');
+$routes->get('/jenisbarang', 'Kategori::index', ['filter' => 'auth']);
+$routes->post('/jenisbarang', 'Kategori::store', ['filter' => 'auth']);
+$routes->get('/jenisbarang/edit/(:any)', 'Kategori::edit/$1', ['filter' => 'auth']);
+$routes->put('/jenisbarang/(:any)', 'Kategori::update/$1', ['filter' => 'auth']);
+$routes->get('/jenisbarang/create', 'Kategori::create', ['filter' => 'auth']);
+$routes->delete('/jenisbarang/(:any)', 'Kategori::destroy/$1', ['filter' => 'auth']);
 
 // Barang
-$routes->get('/barang', 'Barang::index');
-$routes->get('/barang/create', 'Barang::create');
-$routes->put('/barang/(:any)', 'Barang::update/$1');
-$routes->get('/barang/edit/(:any)', 'Barang::edit/$1');
-$routes->post('/barang/save', 'Barang::save');
-$routes->delete('/barang/(:any)', 'Barang::destroy/$1');
+$routes->get('/barang', 'Barang::index', ['filter' => 'auth']);
+$routes->get('/barang/create', 'Barang::create', ['filter' => 'auth']);
+$routes->put('/barang/(:any)', 'Barang::update/$1', ['filter' => 'auth']);
+$routes->get('/barang/edit/(:any)', 'Barang::edit/$1', ['filter' => 'auth']);
+$routes->post('/barang/save', 'Barang::save', ['filter' => 'auth']);
+$routes->delete('/barang/(:any)', 'Barang::destroy/$1', ['filter' => 'auth']);
 
 // Barang Masuk
-$routes->get('/barang_masuk', 'BarangMasuk::index');
-$routes->get('/barang_masuk/create', 'BarangMasuk::create');
-$routes->post('/barang_masuk/save', 'BarangMasuk::save');
-$routes->delete('/barang_masuk/delete/(:num)', 'BarangMasuk::delete/$1');
+$routes->get('/barang_masuk', 'BarangMasuk::index', ['filter' => 'auth']);
+$routes->get('/barang_masuk/create', 'BarangMasuk::create', ['filter' => 'auth']);
+$routes->post('/barang_masuk/save', 'BarangMasuk::save', ['filter' => 'auth']);
+$routes->delete('/barang_masuk/delete/(:num)', 'BarangMasuk::delete/$1', ['filter' => 'auth']);
 
 // Barang Keluar
-$routes->get('/barang_keluar', 'BarangKeluar::index');
-$routes->get('/barang_keluar/create', 'BarangKeluar::create');
-$routes->post('/barang_keluar/save', 'BarangKeluar::save');
-$routes->delete('/barang_keluar/delete/(:num)', 'BarangKeluar::delete/$1');
+$routes->get('/barang_keluar', 'BarangKeluar::index', ['filter' => 'auth']);
+$routes->get('/barang_keluar/create', 'BarangKeluar::create', ['filter' => 'auth']);
+$routes->post('/barang_keluar/save', 'BarangKeluar::save', ['filter' => 'auth']);
+$routes->delete('/barang_keluar/delete/(:num)', 'BarangKeluar::delete/$1', ['filter' => 'auth']);
 
 
 // Supplier
-$routes->get('/data_supplier', 'Supplier::index');
-$routes->get('/data_supplier/create', 'Supplier::create');
-$routes->post('/data_supplier/save', 'Supplier::save');
-$routes->get('/data_supplier/edit/(:num)', 'Supplier::edit/$1');
-$routes->post('/data_supplier/update/(:num)', 'Supplier::update/$1');
-$routes->delete('/data_supplier/delete/(:num)', 'Supplier::delete/$1');
+$routes->get('/data_supplier', 'Supplier::index', ['filter' => 'auth']);
+$routes->get('/data_supplier/create', 'Supplier::create', ['filter' => 'auth']);
+$routes->post('/data_supplier/save', 'Supplier::save', ['filter' => 'auth']);
+$routes->get('/data_supplier/edit/(:num)', 'Supplier::edit/$1', ['filter' => 'auth']);
+$routes->post('/data_supplier/update/(:num)', 'Supplier::update/$1', ['filter' => 'auth']);
+$routes->delete('/data_supplier/delete/(:num)', 'Supplier::delete/$1', ['filter' => 'auth']);
